slight indentation fix
if's and else's with same indentation does not work on Linux. also,
weird thing: left arrows are recognized as right buttons (d's) and up
arrows are recognized as left buttons (a's). works on both Linux and
Mac now, cannot test on Windows because of ANSI, even with ANSICON
(Google it).